+++ src/rezbuild/bin_utils.py
@@ -12,36 +12,8 @@ import struct
 import subprocess
 
 # Import local modules
-from rezbuild.constants import MACHO_MAGIC_NUMBERS
 from rezbuild.exceptions import ArgumentError
 from rezbuild.utils import get_relative_path
-
-
-def add_rpath(macho, rpath):
-    """Add rpath into Macho-O file.
-
-    Only work in macOS.
-
-    Args:
-        macho (str): The path of the macho file to add rpath.
-        rpath (str): The rpath to add.
-    """
-    cmd = ["install_name_tool", "-add_rpath", rpath, macho]
-    subprocess.run(cmd, check=True)
-
-
-def change_lcload(filepath, old_path, new_path):
-    """Change the lcload path.
-
-    Only work in macOS.
-
-    Args:
-        filepath (str): The path of the file to change.
-        old_path (str): The old lcload path to change.
-        new_path (str): The new lcload path to change to.
-    """
-    cmd = ["install_name_tool", "-change", old_path, new_path, filepath]
-    subprocess.run(cmd, check=True)
 
 
 def change_shebang(filepath, shebang, is_bin=False, origin_shebang=""):
@@ -80,41 +52,6 @@ def change_shebang(filepath, shebang, is_bin=False, origin_shebang=""):
         file_.write(new_content)
 
 
-def get_lcload_dylibs(macho, ignore_system_dylib=True):
-    """Get lcload dylibs from Mach-O file.
-
-    Args:
-        macho (str): The Mach-O filepath to get from.
-        ignore_system_dylib (bool): Whether to ignore the system dylibs. System
-            dylibs means the dylib file under `/usr/lib` or under
-            `/System/Library`.
-
-    Returns:
-        :obj:`list` of :obj:`str`: The lcload dylibs.
-    """
-    dylibs = []
-    with open(macho, "rb") as file_:
-        file_.read(16)
-        load_cmd_count = struct.unpack("I", file_.read(4))[0]
-        file_.read(12)
-        for _ in range(load_cmd_count):
-            cmd, size = struct.unpack("2I", file_.read(8))
-            if cmd != 12:
-                file_.read(size - 8)
-                continue
-            file_.read(16)
-            remain_size = size - 8 - 16
-            path = b"".join(struct.unpack(f"{remain_size}c",
-                                          file_.read(remain_size)))
-            path = path.decode("utf-8").strip(b"\x00".decode())
-            if not ignore_system_dylib:
-                dylibs.append(path)
-            elif (not path.startswith("/usr/lib/") and
-                  not path.startswith("/System/Library/")):
-                dylibs.append(path)
-    return dylibs
-
-
 def get_windows_shebang(filepath, pattern):
     """Get the windows shebang from binary files.
 
@@ -135,29 +72,21 @@ def get_windows_shebang(filepath, pattern):
     return ""
 
 
-def is_macho(filepath):
-    """Check if the file is a Mach-O file.
-
-    Args:
-        filepath (str): The path of the file to check.
-
-    Returns:
-        bool: True if the file is a Mach-O file, false otherwise.
-    """
-    with open(filepath, "rb") as f:
-        magic_number = f.read(4)
-    return magic_number in MACHO_MAGIC_NUMBERS
-
-
-def make_bins_movable(bin_path, shebang="", pattern="", rpath="", lib_dir=""):
+def make_bins_movable(
+        bin_path, shebang="", pattern="", rpath="", lib_dir="",
+        change_load_dylib=True):
     """Change all the bins in the directory to make it movable.
 
     Args:
         bin_path (str): Where the entry files placed.
-        shebang (str, optional): The shebang to change to.
-        pattern (str, optional): The re pattern to match shebang.
-        rpath (str, optional): The rpath to add.
+        shebang (str, optional): The shebang to change to. Default is `""`.
+        pattern (str, optional): The re pattern to match shebang. Default is
+            `""`.
+        rpath (str, optional): The rpath to add. Default is `""`.
         lib_dir (str, optional): The destination path to put the dylib file to.
+            Default is `""`.
+        change_load_dylib (str, optional): Whether to change the load dylib
+            path to rpath. Default is True.
     """
     for filename in os.listdir(bin_path):
         filepath = os.path.join(bin_path, filename)
@@ -172,46 +101,205 @@ def make_bins_movable(bin_path, shebang="", pattern="", rpath="", lib_dir=""):
             change_shebang(
                 filepath, shebang, is_bin=True, origin_shebang=origin_shebang)
         else:
-            if is_macho(filepath):
-                make_macho_movable(filepath, rpath, lib_dir)
+            if MachO.is_macho(filepath):
+                MachO(filepath).make_macho_movable(
+                    rpath, lib_dir, change_load_dylib)
             else:
                 change_shebang(filepath, shebang)
 
 
-def make_macho_movable(macho_path, rpath="", lib_dir=""):
-    """Make the macho file movable.
+class MachO(object):
 
-    Copy the unmovable dylibs into the dst_dir and add the rpath into macho.
+    # Mach-O files will contain on of the follow magic numbers at the beginning
+    # of the file.
+    LOAD_COMMANDS_MAPPING = {
+        12: "parse_load_dylib",
+        2147483676: "parse_rpath",
+    }
 
-    Args:
-        macho_path (str): The macho file path to modify.
-        rpath (str, optional): The rpath to add. Default is the relative path
-            between the macho_path and dst_dir.
-        lib_dir (str, optional): The destination path to put the dylib file to.
-            Default is the "lib" folder in the same level directory as the
-            parent directory of macho_path.
-    """
+    # The magic numbers of Mach-O files
+    MACHO_MAGIC_NUMBERS = [
+        b"\xcf\xfa\xed\xfe",
+        b"\xce\xfa\xed\xfe",
+        b"\xca\xfe\xba\xbe",
+    ]
 
-    lib_dir = lib_dir or os.path.join(
-        os.path.dirname(os.path.dirname(macho_path)), "lib")
-    rpath = rpath or os.path.join(
-        "@executable_path",
-        get_relative_path(os.path.dirname(macho_path), lib_dir))
-    add_rpath(macho_path, rpath)
+    # File path prefix of the system dylib.
+    SYSTEM_DYLIB_PREFIX = [
+        "/usr/lib/",
+        "/System/Library/",
+    ]
 
-    def copy_and_change(macho):
-        """Copy dylibs and change lcload path.
+    def __init__(self, path):
+        """Initialize.
 
         Args:
-            macho (str): The macho file to change.
+            path (str): The path of the Mach-O file.
         """
-        dylibs = get_lcload_dylibs(macho)
-        for dylib in dylibs:
-            lib_name = os.path.basename(dylib)
-            change_lcload(macho, dylib, f"@rpath/{lib_name}")
-            if lib_name not in os.listdir(lib_dir):
-                dst = os.path.join(lib_dir, lib_name)
-                shutil.copy2(dylib, dst)
-                copy_and_change(dylib)
+        self.load_dylibs = []
+        self.path = path
+        self.rpaths = []
+        self.parse_load_command()
 
-    copy_and_change(macho_path)
+    def add_rpath(self, rpath):
+        """Add rpath into Macho-O file.
+
+        Args:
+            rpath (str): The rpath to add.
+        """
+        cmd = ["install_name_tool", "-add_rpath", rpath, self.path]
+        subprocess.run(cmd, check=True)
+
+    def change_load_dylib(self, old_path, new_path):
+        """Change the load dylib path.
+
+        Args:
+            old_path (str): The old load dylib to change.
+            new_path (str): The new load dylib to change to.
+        """
+        cmd = ["install_name_tool", "-change", old_path, new_path, self.path]
+        subprocess.run(cmd, check=True)
+
+    @staticmethod
+    def decode_str(content):
+        """Decode bytes to str.
+
+        Args:
+            content (bytes): The bytes to decode.
+        """
+        result = b"".join(struct.unpack(f"{len(content)}c", content))
+        return result.decode("utf-8").strip(b"\x00".decode())
+
+    def get_default_libdir(self):
+        """Get the path of the lib dir under the package root.
+
+        Assume the package tree like this:
+        package_root/
+            |___bin/
+                |___Mach-O file
+            |___include/
+            |___lib/
+                |___lib1
+                |___lib2
+            |___share/
+        """
+        return os.path.join(
+            os.path.dirname(os.path.dirname(self.path)), "lib")
+
+    def get_default_rpath(self, libdir=None):
+        """Get the the rpath relative with the lib path.
+
+        Args:
+            libdir (str, optional): The lib dir to calculate the rpath. Default
+                is the lib dir under the package root.
+        """
+        libdir = libdir or self.get_default_libdir()
+        return os.path.join(
+            "@executable_path",
+            get_relative_path(os.path.dirname(self.path), libdir))
+
+    def get_load_dylibs(self, ignore_system_dylib=False):
+        """Get all the dylibs in this Mach-O files.
+
+        Args:
+            ignore_system_dylib (bool, optional): Whether include the system
+                dylibs. Default is False.
+        """
+        if ignore_system_dylib:
+            return self.get_non_system_dylibs()
+        else:
+            return self.load_dylibs
+
+    def get_non_system_dylibs(self):
+        """Get dylibs that doesn't contain system dylibs."""
+        non_system_dylibs = []
+        for dylib in self.load_dylibs:
+            for prefix in self.SYSTEM_DYLIB_PREFIX:
+                if dylib.startswith(prefix):
+                    break
+            else:
+                non_system_dylibs.append(dylib)
+        return non_system_dylibs
+
+    @classmethod
+    def is_macho(cls, filepath):
+        """Check if the file is a Mach-O file.
+
+        Args:
+            filepath (str): The path of the file to check.
+
+        Returns:
+            bool: True if the file is a Mach-O file, false otherwise.
+        """
+        with open(filepath, "rb") as f:
+            magic_number = f.read(4)
+        return magic_number in cls.MACHO_MAGIC_NUMBERS
+
+    def make_macho_movable(self, rpath="", lib_dir="", change_load_dylib=True):
+        """Make the macho file movable.
+
+        Copy the unmovable dylibs into the dst_dir and add the rpath into
+        macho.
+
+        Args:
+            rpath (str, optional): The rpath to add. Default is the relative
+                path between the macho_path and dst_dir.
+            lib_dir (str, optional): The destination path to put the dylib file
+                to. Default is the "lib" folder in the same level directory as
+                the parent directory of macho_path.
+            change_load_dylib (bool, optional): Whether to change the load
+                dylib path to rpath. Default is True.
+        """
+        lib_dir = lib_dir or self.get_default_libdir()
+        rpath = rpath or self.get_default_rpath(lib_dir)
+        if change_load_dylib and rpath not in self.rpaths:
+            self.add_rpath(rpath)
+
+        def copy_and_change(macho):
+            """Copy dylibs and change lcload path.
+
+            Args:
+                macho (MachO): The macho object to change.
+            """
+            for dylib in macho.get_load_dylibs(ignore_system_dylib=True):
+                lib_name = os.path.basename(dylib)
+                if change_load_dylib:
+                    macho.change_load_dylib(dylib, f"@rpath/{lib_name}")
+                if lib_name not in os.listdir(lib_dir):
+                    dst = os.path.join(lib_dir, lib_name)
+                    shutil.copy2(dylib, dst)
+                    copy_and_change(MachO(dylib))
+
+        copy_and_change(self)
+
+    def parse_load_command(self):
+        """Parse Mach-O file load commands."""
+        with open(self.path, "rb") as file_:
+            file_.read(16)
+            load_cmd_count = struct.unpack("I", file_.read(4))[0]
+            file_.read(12)
+            for _ in range(load_cmd_count):
+                cmd, size = struct.unpack("2I", file_.read(8))
+                content = file_.read(size - 8)
+                if cmd not in self.LOAD_COMMANDS_MAPPING.keys():
+                    continue
+                func = self.LOAD_COMMANDS_MAPPING[cmd]
+                getattr(self, func)(content)
+
+    def parse_rpath(self, content):
+        """Parse rpath from the given rpath content.
+
+        Args:
+            content (bytes): Part of the Mach-O file content that contain the
+                rpath to parse.
+        """
+        self.rpaths.append(self.decode_str(content[4:]))
+
+    def parse_load_dylib(self, content):
+        """Parse load dylib from the given content.
+
+        Args:
+            content (bytes): Part of the Mach-O file content that contain the
+                load dylib to parse.
+        """
+        self.load_dylibs.append(self.decode_str(content[16:]))