+++ src/rezbuild/constants.py
@@ -1,7 +1,0 @@
-"""Constants module."""
-
-MACHO_MAGIC_NUMBERS = [
-    b"\xcf\xfa\xed\xfe",
-    b"\xce\xfa\xed\xfe",
-    b"\xca\xfe\xba\xbe",
-]